+++ src/Blueprint/LaravelMySqlConfiguration.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LaravelCloudBlueprint\Blueprint;
+
+final readonly class LaravelMySqlConfiguration implements DatabaseClusterConfiguration
+{
+    public function __construct(
+        public string $size,
+        public int $storage,
+        public int $retentionDays,
+        public bool $usesScheduledSnapshots,
+        public bool $isPublic,
+    ) {
+    }
+}